+++ src/clients/client.h
@@ -78,9 +78,19 @@ public:
         return m_key;
     }
 
+    inline const std::string& backupKey() const
+    {
+        return m_backupKey;
+    }
+
     inline int keySize() const
     {
         return m_keySize;
+    }
+
+    inline int backupKeySize() const
+    {
+        return m_backupKeySize;
     }
 
     inline void setAge(unsigned int age)
@@ -128,9 +138,19 @@ public:
         m_keySize = keySize;
     }
 
+    inline void setBackupKeySize(int keySize)
+    {
+        m_backupKeySize = keySize;
+    }
+
     inline void setKey(const std::string& key)
     {
         m_key = key;
+    }
+
+    inline void setBackupKey(const std::string& key)
+    {
+        m_backupKey = key;
     }
 
     inline std::unordered_map<std::string, std::unordered_set<Token>>& tokens()
@@ -141,7 +161,7 @@ public:
     bool isAlive(const types::Time& compareTo = 0L) const;
 
     void addToken(const std::string&, const Token& token);
-    void removeToken(const std::string&, const std::string& token);
+    //void removeToken(const std::string&, const std::string& token);
     bool isValidToken(const std::string&,
                       const std::string&,
                       const Registry*,
@@ -159,6 +179,11 @@ private:
     bool m_isKnown;
 
     std::unordered_map<std::string, std::unordered_set<Token>> m_tokens;
+
+    // a backup key is previously set key with potentially different key size
+    // see https://github.com/muflihun/residue/issues/75
+    std::string m_backupKey;
+    int m_backupKeySize;
 };
 }