+++ src/logging/logging-queue.cc
@@ -33,9 +33,7 @@ void LoggingQueue::switchContext()
 {
     std::lock_guard<std::mutex> lock(m_mutex);
 
-    RESIDUE_UNUSED(lock);
-
-#if RESIDUE_DEBUG
+#ifdef RESIDUE_DEBUG
     DRVLOG_IF(!m_queue1.empty() || !m_queue2.empty(), RV_DEBUG)
                     << "Context switched, queue 1: " << m_queue1.size()
                     << " items, queue 2: " << m_queue2.size() << " items";
@@ -48,5 +46,5 @@ RawRequest LoggingQueue::pull()
 {
     RawRequest rawRequest = m_dispatchQueue->back();
     m_dispatchQueue->pop_back();
-    return rawRequest;
+    return std::move(rawRequest);
 }