+++ src/net/session.cc
@@ -46,6 +46,7 @@ Session::Session(tcp::socket&& socket,
 {
     m_id = m_requestHandler->name()[0] + Utils::generateRandomString(16, true);
     DRVLOG(RV_DEBUG) << "New session " << m_id;
+    m_requestHandler->setSession(this);
 }
 
 Session::~Session()
@@ -70,7 +71,7 @@ void Session::read()
 #endif
         if (!ec) {
             RESIDUE_PROFILE_START(t_read);
-#if RESIDUE_DEBUG
+#ifdef RESIDUE_DEV
             DRVLOG(RV_TRACE) << "Received: "
                              << (numOfBytes - Session::PACKET_DELIMITER_SIZE) // ignore package delimiter
                              << " bytes";
@@ -81,10 +82,13 @@ void Session::read()
             //RESIDUE_PROFILE_CHECKPOINT(t_read, m_timeTaken, 1);
             sendToHandler(std::move(buffer));
             //RESIDUE_PROFILE_CHECKPOINT(t_read, m_timeTaken, 2);
+#ifdef RESIDUE_DEV
+            DRVLOG(RV_TRACE) << "Adding bytes";
+#endif
             Utils::bigAdd(m_bytesReceived, std::to_string(numOfBytes));
             m_requestHandler->registry()->addBytesReceived(numOfBytes);
         } else {
-#if RESIDUE_DEBUG
+#ifdef RESIDUE_DEBUG
             DRVLOG_IF(ec != net::error::eof, RV_DEBUG) << "Error: " << ec.message();
 #endif
             m_requestHandler->registry()->leave(shared_from_this());
@@ -94,10 +98,9 @@ void Session::read()
 
 void Session::sendToHandler(std::string&& data)
 {
-#if RESIDUE_DEBUG
+#ifdef RESIDUE_DEBUG
     DRVLOG(RV_TRACE) << "Read bytes: " << data << " [size: " << data.size() << "]";
 #endif
-    m_requestHandler->setSession(this);
     RawRequest req { std::move(data), m_socket.remote_endpoint().address().to_string(), Utils::now() };
     m_requestHandler->handle(std::move(req));
 }
@@ -105,7 +108,7 @@ void Session::sendToHandler(std::string&& data)
 void Session::close()
 {
     if (m_socket.is_open()) {
-#if RESIDUE_DEBUG
+#ifdef RESIDUE_DEBUG
         DRVLOG(RV_DEBUG) << "Closing session...";
 #endif
         m_socket.close();
@@ -164,13 +167,13 @@ void Session::write(const char* data,
     Utils::bigAdd(m_bytesSent, std::to_string(length));
     m_requestHandler->registry()->addBytesSent(length);
 
-#if RESIDUE_DEBUG
-    DRVLOG(RV_DEBUG) << "Sending " << data;
+#ifdef RESIDUE_DEBUG
+    DRVLOG(RV_DEBUG_2) << "Sending " << data;
 #endif
     net::async_write(m_socket, net::buffer(data, length),
                      [&, this, self](residue::error_code ec, std::size_t) {
         if (ec) {
-#if RESIDUE_DEBUG
+#ifdef RESIDUE_DEBUG
             DRVLOG(RV_DEBUG) << "Failed to send." << ec.message();
 #endif
             m_socket.close();